Refactor methods to top level functions
